creates a homepage
